refactor install.sh: add checks for existing scripts before downloading

+++ installer/install.sh
@@ -2,12 +2,24 @@
 
 serverport=$1
 
-wget -q https://github.com/soumas/model-flight-logbook/raw/refs/heads/main/installer/install-server.sh
+rmfile=false
+if [ ! -f "install-server.sh" ]; then
+    wget -q https://github.com/soumas/model-flight-logbook/raw/refs/heads/main/installer/install-server.sh
+    rmfile=true
+fi
 sudo chmod +x ./install-server.sh
 ./install-server.sh $serverport
-rm ./install-server.sh
+if [ "$rmfile" = true ]; then
+    rm ./install-server.sh
+fi
 
-wget -q https://github.com/soumas/model-flight-logbook/raw/refs/heads/main/installer/install-terminal.sh
+rmfile=false
+if [ ! -f "install-terminal.sh" ]; then
+    wget -q https://github.com/soumas/model-flight-logbook/raw/refs/heads/main/installer/install-terminal.sh
+    rmfile=true
+fi
 sudo chmod +x ./install-terminal.sh
 ./install-terminal.sh $serverport
-rm ./install-terminal.sh
+if [ "$rmfile" = true ]; then
+    rm ./install-terminal.sh
+fi
